# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Messenger",
        "description": "Мессенджер",
        "version": "1"
    },
    "paths": {
        "/chat": {
            "get": {
                "tags": [
                    "chat"
                ],
                "summary": "Получение списка чатов пользователя",
                "description": "Получение списка чатов пользователя",
                "operationId": "cb94f170cf02ce390775ec8ab58363e6",
                "parameters": [
                    {
                        "parameter": "page",
                        "name": "page",
                        "in": "query",
                        "description": "Текущая страница",
                        "required": false,
                        "deprecated": false,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Successfully"
                                        },
                                        "paginator": {
                                            "$ref": "#/components/schemas/Paginator"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "1"
                                                    },
                                                    "user_id1": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "23"
                                                    },
                                                    "user_id2": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "11"
                                                    },
                                                    "created_at": {
                                                        "type": "TIMESTAMP",
                                                        "format": "TIMESTAMP",
                                                        "example": "2025-01-19 19:55:24"
                                                    },
                                                    "user1": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "format": "int",
                                                                    "example": "1"
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "format": "name",
                                                                    "example": "Oleg"
                                                                },
                                                                "email": {
                                                                    "type": "string",
                                                                    "format": "email",
                                                                    "example": "11@mail.ru"
                                                                },
                                                                "created_at": {
                                                                    "type": "TIMESTAMP",
                                                                    "format": "TIMESTAMP",
                                                                    "example": "2025-01-19 19:55:24"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "user2": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "format": "int",
                                                                    "example": "1"
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "format": "name",
                                                                    "example": "Oleg"
                                                                },
                                                                "email": {
                                                                    "type": "string",
                                                                    "format": "email",
                                                                    "example": "11@mail.ru"
                                                                },
                                                                "created_at": {
                                                                    "type": "TIMESTAMP",
                                                                    "format": "TIMESTAMP",
                                                                    "example": "2025-01-19 19:55:24"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "latest_message": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "format": "int",
                                                                    "example": "1"
                                                                },
                                                                "message": {
                                                                    "type": "string",
                                                                    "format": "string",
                                                                    "example": "shsga4rg34"
                                                                },
                                                                "user_id": {
                                                                    "type": "integer",
                                                                    "format": "int",
                                                                    "example": "2"
                                                                },
                                                                "chat_id": {
                                                                    "type": "integer",
                                                                    "format": "int",
                                                                    "example": "22"
                                                                },
                                                                "viewed_at": {
                                                                    "type": "TIMESTAMP",
                                                                    "format": "TIMESTAMP",
                                                                    "example": "2025-01-19 19:55:24"
                                                                },
                                                                "created_at": {
                                                                    "type": "TIMESTAMP",
                                                                    "format": "TIMESTAMP",
                                                                    "example": "2025-01-19 19:55:24"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "chat"
                ],
                "summary": "Создание чата для пользователя",
                "description": "Создание чата для пользователя",
                "operationId": "9ae167ebeb88c5ed41247d0af122b71d",
                "requestBody": {
                    "$ref": "#/components/requestBodies/ChatCreateRequest"
                },
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Successfully"
                                        },
                                        "paginator": {
                                            "$ref": "#/components/schemas/Paginator"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "1"
                                                    },
                                                    "user_id1": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "23"
                                                    },
                                                    "user_id2": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "11"
                                                    },
                                                    "created_at": {
                                                        "type": "TIMESTAMP",
                                                        "format": "TIMESTAMP",
                                                        "example": "2025-01-19 19:55:24"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/chat/{id}": {
            "get": {
                "tags": [
                    "chat"
                ],
                "summary": "Получение одного чата пользователя",
                "description": "Получение одного чата пользователя",
                "operationId": "8c680e539b37a9f303eb82fa816ac714",
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Successfully"
                                        },
                                        "paginator": {
                                            "$ref": "#/components/schemas/Paginator"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "1"
                                                    },
                                                    "user_id1": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "23"
                                                    },
                                                    "user_id2": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "11"
                                                    },
                                                    "created_at": {
                                                        "type": "TIMESTAMP",
                                                        "format": "TIMESTAMP",
                                                        "example": "2025-01-19 19:55:24"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "chat"
                ],
                "summary": "Удаление чата пользователя",
                "description": "Удаление чата пользователя",
                "operationId": "9b2192d76ea2b33383d0328e42ff9ada",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "id",
                        "in": "query",
                        "description": "Идентификатор чата",
                        "required": true,
                        "deprecated": false,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Paginator": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "format": "int",
                        "example": "1"
                    },
                    "first_page_url": {
                        "type": "string",
                        "format": "url",
                        "example": "http://138.124.55.208/api/prod/chat?page=1"
                    },
                    "from": {
                        "type": "integer",
                        "format": "int",
                        "example": "1"
                    },
                    "last_page": {
                        "type": "integer",
                        "format": "int",
                        "example": "380"
                    },
                    "last_page_url": {
                        "type": "string",
                        "format": "url",
                        "example": "http://138.124.55.208/api/prod/chat?page=5"
                    },
                    "next_page_url": {
                        "type": "string",
                        "format": "url",
                        "example": "http://138.124.55.208/api/prod/chat?page=3"
                    },
                    "path": {
                        "type": "string",
                        "format": "url",
                        "example": "http://138.124.55.208/api/prod/chat?page=2"
                    },
                    "per_page": {
                        "type": "integer",
                        "format": "int",
                        "example": "30"
                    },
                    "to": {
                        "type": "integer",
                        "format": "int",
                        "example": "30"
                    },
                    "total": {
                        "type": "integer",
                        "format": "int",
                        "example": "11371"
                    }
                },
                "type": "object"
            }
        },
        "requestBodies": {
            "ChatCreateRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "user_id2": {
                                    "schema": "required|int",
                                    "type": "integer",
                                    "format": "int",
                                    "example": "4"
                                },
                                "message": {
                                    "schema": "required|string",
                                    "type": "string",
                                    "format": "string",
                                    "example": "dssvsdvsv"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        }
    },
...
}